Moved toggle start stop logic into Service
