+++ gpslogger/src/main/java/com/mendhak/gpslogger/common/events/ServiceEvents.java
@@ -60,4 +60,7 @@ public class ServiceEvents {
             this.newFileName = newFileName;
         }
     }
+
+    public static class RequestToggleEvent {
+    }
 }
