+++ gpslogger/src/main/java/com/mendhak/gpslogger/views/GpsBigViewFragment.java
@@ -107,8 +107,7 @@ public class GpsBigViewFragment extends GenericViewFragment implements View.OnTo
     @Override
     public boolean onTouch(View view, MotionEvent motionEvent) {
         if (motionEvent.getAction() == MotionEvent.ACTION_DOWN) {
-            tracer.debug("Big frame - onTouch event");
-            requestToggleLogging();
+            RequestToggleLogging();
             return true;
         }
